+++ tools/openfeature-flagd-core/src/openfeature/contrib/tools/flagd/core/targeting/__init__.py
@@ -1,0 +1,3 @@
+from .targeting import targeting
+
+__all__ = ["targeting"]