+++ tools/openfeature-flagd-core/tests/e2e/test_evaluator.py
@@ -1,0 +1,5 @@
+from pytest_bdd import scenarios
+
+from openfeature.contrib.tools.flagd.testkit import get_features_path
+
+scenarios(get_features_path())